requirements.txt
# Core dependencies
numpy>=1.24.0
pygame>=2.5.0

# GPU Acceleration
numba>=0.57.0
cuda-python>=12.0.0

# Image Processing
Pillow>=10.0.0  # This is the PIL library